integracionKafkaConsumer2/requirements.txt
python-dotenv
confluent-kafka
pymongo